lib: added USB error handling callback to SetUsbDevice()
Since there is only one USB device per routing queue, allow the
application to define one central place for code to recover from
potential USB errors.

Most USB errors are catastrophic, and need a full Probe and Controller
to recover, but it should be possible to preserve the routing queue
through it all.